# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MathMind AI EdTech Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered user with token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Authenticated user with token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /battles:
    get:
      operationId: listBattles
      tags: [battles]
      summary: List open battle rooms
      responses:
        "200":
          description: Open rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Battle"
    post:
      operationId: createBattle
      tags: [battles]
      summary: Create a new battle room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBattleBody"
      responses:
        "201":
          description: Created battle room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Battle"
  /battles/{id}:
    get:
      operationId: getBattle
      tags: [battles]
      summary: Get battle room details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Battle room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BattleWithParticipants"
  /battles/{id}/join:
    post:
      operationId: joinBattle
      tags: [battles]
      summary: Join a battle room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined battle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BattleWithParticipants"
  /battles/{id}/submit:
    post:
      operationId: submitBattleAnswer
      tags: [battles]
      summary: Submit answer in a battle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BattleAnswerBody"
      responses:
        "200":
          description: Answer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BattleAnswerResult"

  # Friends
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # Auth
    RegisterBody:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        displayName:
          type: string
        language:
          type: string
          enum: [en, hi]
      required:
        - username
        - email
        - password
        - displayName
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user

    # Users
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        language:
          type: string
          enum: [en, hi]
        xp:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        totalQuestionsAnswered:
          type: integer
        totalCorrect:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - displayName
        - language
        - xp
        - level
        - streak
        - totalQuestionsAnswered
        - totalCorrect
        - createdAt
...
    Question:
      type: object
      properties:
        id:
          type: integer
        topicId:
          type: integer
        text:
          type: string
        textHi:
          type: ["string", "null"]
        options:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        points:
          type: integer
        timeLimit:
          type: integer
      required:
        - id
        - topicId
        - text
        - options
        - difficulty
        - points
        - timeLimit
...
    Battle:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        subjectId:
          type: integer
        subjectName:
          type: string
        hostId:
          type: integer
        hostName:
          type: string
        status:
          type: string
          enum: [waiting, in_progress, finished]
        maxParticipants:
          type: integer
        currentParticipants:
          type: integer
        questionCount:
          type: integer
        timePerQuestion:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - code
        - subjectId
        - subjectName
        - hostId
        - hostName
        - status
        - maxParticipants
        - currentParticipants
        - questionCount
        - timePerQuestion
        - createdAt
    BattleParticipant:
      type: object
      properties:
        userId:
          type: integer
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        score:
          type: integer
        correctAnswers:
          type: integer
        isReady:
          type: boolean
      required:
        - userId
        - displayName
        - score
        - correctAnswers
        - isReady
    BattleWithParticipants:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        subjectId:
          type: integer
        subjectName:
          type: string
        hostId:
          type: integer
        hostName:
          type: string
        status:
          type: string
          enum: [waiting, in_progress, finished]
        maxParticipants:
          type: integer
        questionCount:
          type: integer
        timePerQuestion:
          type: integer
        participants:
          type: array
          items:
            $ref: "#/components/schemas/BattleParticipant"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - code
        - subjectId
        - subjectName
        - hostId
        - hostName
        - status
        - maxParticipants
        - questionCount
        - timePerQuestion
        - participants
        - questions
        - createdAt
    CreateBattleBody:
      type: object
      properties:
        subjectId:
          type: integer
        questionCount:
          type: integer
        timePerQuestion:
          type: integer
      required:
        - subjectId
    BattleAnswerBody:
      type: object
      properties:
        questionId:
          type: integer
        answer:
          type: string
        timeTaken:
          type: integer
      required:
        - questionId
        - answer
        - timeTaken
    BattleAnswerResult:
      type: object
      properties:
        correct:
          type: boolean
        correctAnswer:
          type: string
        pointsEarned:
          type: integer
        currentScore:
          type: integer
      required:
        - correct
        - correctAnswer
        - pointsEarned
        - currentScore

    # Friends
...